# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"fullname": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"current_password": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"ACTIVE",
							"INACTIVE",
							"PENDING"
						]
					},
					"resetPasswordToken": {
						"type": "string",
						"nullable": true
					},
					"verificationCode": {
						"type": "string",
						"nullable": true
					},
					"verificationCodeExpires": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"twoFactorCode": {
						"type": "string",
						"nullable": true
					},
					"twoFactorExpires": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"created_at",
					"updated_at",
					"fullname",
					"phone",
					"email",
					"current_password",
					"roleId",
					"status",
					"resetPasswordToken",
					"verificationCode",
					"verificationCodeExpires",
					"twoFactorCode",
					"twoFactorExpires"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerificationResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "logistics-app-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"phone": {
										"type": "string"
									},
									"roleId": {
										"type": "string"
									},
									"current_password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"fullname": {
										"type": "string"
									}
								},
								"required": [
									"phone",
									"current_password",
									"email",
									"fullname"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerificationResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerificationResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									}
								},
								"required": [
									"newPassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/verify-email": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerificationResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"code",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/signIn": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerificationResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"current_password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"current_password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/verify-code": {
			"post": {
				"operationId": "VerifyCodeSms",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"code",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}